bugs fixed

+++ frontend/src/api/authApi.js
@@ -14,18 +14,18 @@ const authApi = {
     updatePassword: (data) =>
         axiosClient.post('/auth/update-password', data),
 
-    // Add more:
     login: (data) => axiosClient.post("/auth/login", data),
     signup: (data) => axiosClient.post("/auth/signup", data),
     ssoExchange: (data) => axiosClient.post("/sso-exchange", data),
+
     // Admin APIs
     getAllUsers: () => axiosClient.get("/admin/users"),
-    updateUserStatus: (userId, status) =>
-        axiosClient.put(`/admin/users/${userId}/status`, { status }),
-    deleteUser: (userId) =>
-        axiosClient.delete(`/admin/users/${userId}`),
+    getUserByEmail: (userId) => axiosClient.post(`/admin/get-user`, userId),
 
-    getUserByEmail: (userId) => axiosClient.post(`/admin/get-user`, userId)
+    updateUserStatus: (userId, status) => axiosClient.put(`/admin/users/${userId}/status`, { status }),
+    updateUserRole: (userId, role) => axiosClient.patch(`/admin/update-user-role/${userId}`, { role }),
+    deleteUser: (userId) => axiosClient.delete(`/admin/users/${userId}`),
+    createCompanyUser: (data) => axiosClient.post('/admin/inser-company-user', data),
 };
 
 export default authApi;
